% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Destin2.R
\name{GetConsensusPCA}
\alias{GetConsensusPCA}
\title{Consensus PCA of Seurat}
\usage{
GetConsensusPCA(
  obj,
  reduction.list,
  dims.list,
  assay = NULL,
  reduction.name = NULL,
  reduction.key = NULL,
  n.consensus.pc = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{reduction.list}{A list of of dimension reduction objects within obj}

\item{dims.list}{A list of which dimensions to use for each combined dimension reduction object}

\item{assay}{A character string specifying the assay used to calculate this dimensional reduction}

\item{reduction.name}{A character string providing the name of the CPCA dimension reduction}

\item{reduction.key}{A character string that acts as a prefix for the CPCA dimension reduction}

\item{n.consensus.pc}{Number of principal components (PC's) to store}
}
\value{
A Seurat object containing CPCA joint dimension reduction in slot "consensus.pca"
}
\description{
This function takes a Seurat object and conducts consensus principal component analysis (CPCA) as joint dimension reduction
}
