% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Destin2.R
\name{GetConsensusPCA}
\alias{GetConsensusPCA}
\title{Consensus PCA of Seurat}
\usage{
GetConsensusPCA(
  obj,
  reduction.list,
  dims.list,
  reduction.name = NULL,
  reduction.key = NULL,
  n.consensus.pc = NULL
)
}
\arguments{
\item{obj}{a Seurat object}

\item{reduction.list}{a list of of dimensional reduction objects within obj}

\item{reduction.name}{a character providing the name of the CPCA dimension reduction}

\item{reduction.key}{a character string that acts as a prefix for the CPCA dimension reduction}

\item{n.consensus.pc}{Number of PC's to store}
}
\value{
a Seurat object containing CPCA joint dimension reduction
}
\description{
This function takes a Seurat object and returns consensus PCA (CPCA) as joint dimension reduction
}
