% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Destin2.R
\name{GetMultiCCA}
\alias{GetMultiCCA}
\title{MultiCCA of Seurat}
\usage{
GetMultiCCA(
  obj,
  reduction.list,
  dims.list,
  reduction.name = NULL,
  reduction.key = NULL,
  n.cca = NULL
)
}
\arguments{
\item{obj}{a Seurat object}

\item{reduction.list}{a list of of dimensional reduction objects within obj}

\item{reduction.name}{a character providing the name of the MCCA dimension reduction}

\item{reduction.key}{a character string that acts as a prefix for the MCCA dimension reduction}

\item{n.consensus.pc}{Number of PC's to store}
}
\value{
a Seurat object containing MCCA joint dimension reduction
}
\description{
This function takes a Seurat object and returns multiple canonical correlation analysis (MCCA) as joint dimension reduction
}
