% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Destin2.R
\name{GetSlingshot}
\alias{GetSlingshot}
\title{Get Slingshot Lineages of Seurat object}
\usage{
GetSlingshot(
  obj,
  reduction.to.construct,
  reduction.to.plot,
  cluster,
  predicted.id = NULL,
  generate.plot = NULL
)
}
\arguments{
\item{obj}{a Seurat object}

\item{reduction.to.construct}{a character string specifying dimension reduction object used to infer lineages}

\item{reduction.to.plot}{a character string specfiying dimension reduction object to cluster centroids and plot lineages}

\item{cluster}{a vector specifying cluster identity for each sample}

\item{predicted.id}{a character string specifying (transferred) cell labels}

\item{generate.plot}{a boolean, if true returns Slingshot plots based on cluster and predicted.id}
}
\value{
Slingshot Lineages
}
\description{
This function takes a Seurat object and displays trajectory inference using Slingshot
}
